backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
